+++ src/constants/errors.ts
@@ -41,7 +41,6 @@ export const COURSE_ERRORS = {
   DELETE_DISALLOWED:
     'Digi Dojo does not allow any deletion of published courses',
   LESSON_MISSING: 'Lesson does not exist',
-  NOTE_MISSING: 'Note does not exist',
   ANNOUNCEMENT_MISSING: 'Annoucement does not exist',
   USER_NOT_VERIFIED:
     'Your account has not been verified by Diji Dojo. Please complete your verification first before publishing your course or submitting your course request for approval.',
@@ -68,6 +67,7 @@ export const MENTORSHIP_ERRORS = {
     'Your account has not been verified by Diji Dojo. Please complete your verification first before publishing your mentorship listing.',
   TASK_BUCKET_MISSING: 'Task bucket does not exist',
   TASK_MISSING: 'Task does not exist',
+  NOTE_MISSING: 'Note does not exist',
 };
 
 export const REVIEW_ERRORS = {
