+++ src/constants/successMessages.ts
@@ -45,9 +45,6 @@ export const COURSE_RESPONSE = {
   LESSON_UPDATE: 'Lesson has been successfully updated',
   LESSON_DELETE: 'Lesson has been successfully deleted',
 
-  NOTE_CREATE: 'Note has been successfully created',
-  NOTE_UPDATE: 'Note has been successfully updated',
-
   COURSE_REQUEST_ACCEPTED: 'Course Request has been successfully accepted',
   COURSE_REQUEST_REJECTED: 'Course Request has been successfully rejected',
 
@@ -92,6 +89,9 @@ export const MENTORSHIP_RESPONSE = {
   TASK_CREATE: 'Task has been successfully created',
   TASK_EDIT: 'Task has been successfully edited',
   TASK_DELETE: 'Task has been successfully deleted',
+
+  NOTE_CREATE: 'Note has been successfully created',
+  NOTE_UPDATE: 'Note has been successfully updated',
 };
 
 // ==================== REFUND RESPONSE MESSAGES ====================
