+++ src/controllers/course.controller.ts
@@ -266,112 +266,6 @@ export class CourseController {
       });
     }
   }
-  // ======================================== NOTES ================================================
-  public static async addNoteToLesson(req, res) {
-    const { user } = req;
-    const { lessonId } = req.params;
-    const { newNote } = req.body;
-
-    try {
-      const addedNote = await CourseService.addNoteToLesson(
-        lessonId,
-        user.accountId,
-        newNote
-      );
-      return apiResponse.result(
-        res,
-        {
-          message: COURSE_RESPONSE.NOTE_CREATE,
-          note: addedNote,
-        },
-        httpStatusCodes.CREATED
-      );
-    } catch (e) {
-      logger.error('[courseController.addNoteToLesson]:' + e.message);
-      if (
-        e.message ===
-          httpStatusCodes.getStatusText(httpStatusCodes.UNAUTHORIZED) ||
-        e.message === COURSE_ERRORS.COURSE_MISSING ||
-        e.message === COURSE_ERRORS.LESSON_MISSING
-      ) {
-        return apiResponse.error(res, httpStatusCodes.BAD_REQUEST, {
-          message: e.message,
-        });
-      }
-      return apiResponse.error(res, httpStatusCodes.INTERNAL_SERVER_ERROR, {
-        message: RESPONSE_ERROR.RES_ERROR,
-      });
-    }
-  }
-
-  public static async editNoteInLesson(req, res) {
-    const { user } = req;
-    const { noteId } = req.params;
-    const { updateNote } = req.body;
-    try {
-      const updatedNote = await CourseService.editNoteInLesson(
-        noteId,
-        user.accountId,
-        updateNote
-      );
-      return apiResponse.result(
-        res,
-        {
-          message: COURSE_RESPONSE.NOTE_UPDATE,
-          note: updatedNote,
-        },
-        httpStatusCodes.OK
-      );
-    } catch (e) {
-      logger.error('[courseController.editNoteInLesson]:' + e.message);
-      if (
-        e.message ===
-          httpStatusCodes.getStatusText(httpStatusCodes.UNAUTHORIZED) ||
-        e.message === COURSE_ERRORS.NOTE_MISSING ||
-        e.message === COURSE_ERRORS.COURSE_MISSING ||
-        e.message === COURSE_ERRORS.LESSON_MISSING
-      ) {
-        return apiResponse.error(res, httpStatusCodes.BAD_REQUEST, {
-          message: e.message,
-        });
-      }
-      return apiResponse.error(res, httpStatusCodes.INTERNAL_SERVER_ERROR, {
-        message: RESPONSE_ERROR.RES_ERROR,
-      });
-    }
-  }
-
-  public static async getAllNotes(req, res) {
-    const { lessonId } = req.params;
-    const { user } = req;
-    try {
-      const notes = await CourseService.getAllNotes(lessonId, user.accountId);
-      return apiResponse.result(
-        res,
-        {
-          message: 'success',
-          notes,
-        },
-        httpStatusCodes.OK
-      );
-    } catch (e) {
-      logger.error('[courseController.getAllNotes]:' + e.message);
-      if (
-        e.message ===
-          httpStatusCodes.getStatusText(httpStatusCodes.UNAUTHORIZED) ||
-        e.message === COURSE_ERRORS.COURSE_MISSING ||
-        e.message === COURSE_ERRORS.LESSON_MISSING
-      ) {
-        return apiResponse.error(res, httpStatusCodes.BAD_REQUEST, {
-          message: e.message,
-        });
-      } else {
-        return apiResponse.error(res, httpStatusCodes.INTERNAL_SERVER_ERROR, {
-          message: RESPONSE_ERROR.RES_ERROR,
-        });
-      }
-    }
-  }
 
   // ======================================== ANNOUNCEMENTS ========================================
   public static async createAnnouncement(req, res) {
