+++ src/routes/course.route.ts
@@ -89,33 +89,6 @@ router.delete(
   Utility.asyncHandler(CourseController.deleteLesson)
 );
 
-// ======================================== NOTES ========================================
-router.post(
-  '/note/:lessonId',
-  passport.authenticate('isAuthenticated', { session: false }),
-  requireSensei,
-  schemaValidator.params(course.lessonIdP),
-  schemaValidator.body(course.createNoteB),
-  Utility.asyncHandler(CourseController.addNoteToLesson)
-);
-
-router.put(
-  '/note/:noteId',
-  passport.authenticate('isAuthenticated', { session: false }),
-  requireSensei,
-  schemaValidator.params(course.noteIdP),
-  schemaValidator.body(course.updateNoteB),
-  Utility.asyncHandler(CourseController.editNoteInLesson)
-);
-
-//Get ALL notes added for lesson
-router.get(
-  '/note/all/:lessonId',
-  passport.authenticate('isAuthenticated', { session: false }),
-  schemaValidator.params(course.lessonIdP),
-  Utility.asyncHandler(CourseController.getAllNotes)
-);
-
 // ======================================== ANNOUNCEMENTS ========================================
 router.post(
   '/announcement/:courseId',
