+++ src/routes/schema/course.schema.ts
@@ -14,14 +14,6 @@ export default {
       })
       .required(),
   }),
-  createNoteB: joi.object({
-    newNote: joi
-      .object({
-        title: joi.string().required(),
-        body: joi.string().required(),
-      })
-      .required(),
-  }),
   createCourseB: joi.object({
     newCourse: joi
       .object({
@@ -46,9 +38,6 @@ export default {
   }),
   lessonIdP: joi.object({
     lessonId: joi.string().required(),
-  }),
-  noteIdP: joi.object({
-    noteId: joi.string().required(),
   }),
   announcementIdP: joi.object({
     announcementId: joi.string().required(),
@@ -95,14 +84,6 @@ export default {
       .object({
         title: joi.string().optional(),
         description: joi.string().optional(),
-      })
-      .required(),
-  }),
-  updateNoteB: joi.object({
-    updateNote: joi
-      .object({
-        title: joi.string().optional(),
-        body: joi.string().optional(),
       })
       .required(),
   }),
